unify geo models
