+++ pkg/dss/scd/operations.go
@@ -8,6 +8,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/interuss/dss/pkg/api/v1/scdpb"
 	"github.com/interuss/dss/pkg/dss/auth"
+	dssmodels "github.com/interuss/dss/pkg/dss/models"
 	scdmodels "github.com/interuss/dss/pkg/dss/scd/models"
 	dsserr "github.com/interuss/dss/pkg/errors"
 )
@@ -94,7 +95,7 @@ func (a *Server) SearchOperationReferences(ctx context.Context, req *scdpb.Searc
 	}
 
 	// Parse area of interest to common Volume4D
-	vol4, err := scdmodels.Volume4DFromProto(aoi)
+	vol4, err := dssmodels.Volume4DFromSCDProto(aoi)
 	if err != nil {
 		return nil, dsserr.Internal("failed to convert to internal geometry model")
 	}
@@ -138,7 +139,7 @@ func (a *Server) PutOperationReference(ctx context.Context, req *scdpb.PutOperat
 
 	var (
 		params  = req.GetParams()
-		extents = make([]*scdmodels.Volume4D, len(params.GetExtents()))
+		extents = make([]*dssmodels.Volume4D, len(params.GetExtents()))
 	)
 
 	if len(params.UssBaseUrl) == 0 {
@@ -146,13 +147,13 @@ func (a *Server) PutOperationReference(ctx context.Context, req *scdpb.PutOperat
 	}
 
 	for idx, extent := range params.GetExtents() {
-		cExtent, err := scdmodels.Volume4DFromProto(extent)
+		cExtent, err := dssmodels.Volume4DFromSCDProto(extent)
 		if err != nil {
 			return nil, dsserr.BadRequest(fmt.Sprintf("failed to parse extents: %s", err))
 		}
 		extents[idx] = cExtent
 	}
-	uExtent, err := scdmodels.UnionVolumes4D(extents...)
+	uExtent, err := dssmodels.UnionVolumes4D(extents...)
 	if err != nil {
 		return nil, dsserr.BadRequest(fmt.Sprintf("failed to union extents: %s", err))
 	}
