+++ pkg/dss/scd/store/cockroach/operations.go
@@ -347,13 +347,13 @@ func (s *Store) UpsertOperation(ctx context.Context, operation *scdmodels.Operat
 	// whether all affected OVNs are matched.
 	switch operation.State {
 	case scdmodels.OperationStateAccepted, scdmodels.OperationStateActivated:
-		operations, err := s.searchOperations(ctx, tx, &scdmodels.Volume4D{
+		operations, err := s.searchOperations(ctx, tx, &dssmodels.Volume4D{
 			StartTime: operation.StartTime,
 			EndTime:   operation.EndTime,
-			SpatialVolume: &scdmodels.Volume3D{
+			SpatialVolume: &dssmodels.Volume3D{
 				AltitudeHi: operation.AltitudeUpper,
 				AltitudeLo: operation.AltitudeLower,
-				Footprint: scdmodels.GeometryFunc(func() (s2.CellUnion, error) {
+				Footprint: dssmodels.GeometryFunc(func() (s2.CellUnion, error) {
 					return operation.Cells, nil
 				}),
 			},
@@ -388,7 +388,7 @@ func (s *Store) UpsertOperation(ctx context.Context, operation *scdmodels.Operat
 	return area, subscribers, nil
 }
 
-func (s *Store) searchOperations(ctx context.Context, q dsssql.Queryable, v4d *scdmodels.Volume4D, owner dssmodels.Owner) ([]*scdmodels.Operation, error) {
+func (s *Store) searchOperations(ctx context.Context, q dsssql.Queryable, v4d *dssmodels.Volume4D, owner dssmodels.Owner) ([]*scdmodels.Operation, error) {
 	var (
 		operationsIntersectingVolumeQuery = fmt.Sprintf(`
 			SELECT
@@ -452,7 +452,7 @@ func (s *Store) searchOperations(ctx context.Context, q dsssql.Queryable, v4d *s
 }
 
 // SearchOperations returns operations within the 4D volume from CockroachDB
-func (s *Store) SearchOperations(ctx context.Context, v4d *scdmodels.Volume4D, owner dssmodels.Owner) ([]*scdmodels.Operation, error) {
+func (s *Store) SearchOperations(ctx context.Context, v4d *dssmodels.Volume4D, owner dssmodels.Owner) ([]*scdmodels.Operation, error) {
 	tx, err := s.Begin()
 	if err != nil {
 		return nil, err
