+++ pkg/dss/scd/store/cockroach/subscriptions.go
@@ -229,11 +229,11 @@ func (c *Store) fetchSubscriptionByIDAndOwner(ctx context.Context, q dsssql.Quer
 	if err != nil {
 		return nil, err
 	}
-	ops, err := c.searchOperations(ctx, q, &scdmodels.Volume4D{
-		SpatialVolume: &scdmodels.Volume3D{
+	ops, err := c.searchOperations(ctx, q, &dssmodels.Volume4D{
+		SpatialVolume: &dssmodels.Volume3D{
 			AltitudeLo: result.AltitudeLo,
 			AltitudeHi: result.AltitudeHi,
-			Footprint: scdmodels.GeometryFunc(func() (s2.CellUnion, error) {
+			Footprint: dssmodels.GeometryFunc(func() (s2.CellUnion, error) {
 				return result.Cells, nil
 			}),
 		},
@@ -423,13 +423,13 @@ func (c *Store) UpsertSubscription(ctx context.Context, s *scdmodels.Subscriptio
 
 	var affectedOperations []*scdmodels.Operation
 	if len(s.Cells) > 0 {
-		ops, err := c.searchOperations(ctx, tx, &scdmodels.Volume4D{
+		ops, err := c.searchOperations(ctx, tx, &dssmodels.Volume4D{
 			StartTime: s.StartTime,
 			EndTime:   s.EndTime,
-			SpatialVolume: &scdmodels.Volume3D{
+			SpatialVolume: &dssmodels.Volume3D{
 				AltitudeLo: s.AltitudeLo,
 				AltitudeHi: s.AltitudeHi,
-				Footprint: scdmodels.GeometryFunc(func() (s2.CellUnion, error) {
+				Footprint: dssmodels.GeometryFunc(func() (s2.CellUnion, error) {
 					return s.Cells, nil
 				}),
 			},
@@ -495,7 +495,6 @@ func (c *Store) DeleteSubscription(ctx context.Context, id scdmodels.ID, owner d
 
 	res, err := tx.ExecContext(ctx, query, id, owner)
 	if err != nil {
-		fmt.Println("ARGH:", err)
 		return nil, multierr.Combine(err, tx.Rollback())
 	}
